refactor: move most components back to components/common  & move project to components
