--- src/components/index.ts
@@ -1,4 +1,0 @@
-export { Header } from "components/Header";
-export { ParseHTML } from "components/ParseHTML";
-export { To } from "components/To";
-export { Tooltip } from "components/Tooltip";
