+++ src/components/pages/Error.tsx
@@ -1,4 +1,4 @@
-import { Header } from "components";
+import { Header } from "components/common";
 import { Page } from "components/pages";
 
 import styled from "styled-components";
