+++ src/components/repository/details/Detail.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren } from "react";
+import type { PropsWithChildren } from "react";
 
 import styled from "styled-components";
