+++ src/components/repository/details/Forks.tsx
@@ -1,5 +1,5 @@
 import { ForksIcon } from "commons/icons";
-import Detail from "sections/projects/project/details/Detail";
+import Detail from "components/repository/details/Detail";
 
 export const Forks = ({ forks }: { forks: number }): JSX.Element | null => (
     <Detail detail={forks}>
