+++ src/components/repository/details/Stars.tsx
@@ -1,5 +1,5 @@
 import { StargazersIcon } from "commons/icons";
-import Detail from "sections/projects/project/details/Detail";
+import Detail from "components/repository/details/Detail";
 
 export const Stars = ({ stargazers }: { stargazers: number }): JSX.Element | null => (
     <Detail detail={stargazers}>
