+++ src/components/repository/details/index.ts
@@ -1,0 +1,2 @@
+export { Forks } from "components/repository/details/Forks";
+export { Stars } from "components/repository/details/Stars";
