+++ src/pages/index.tsx
@@ -1,15 +1,15 @@
 import type { GetStaticProps } from "next";
 import { useRef } from "react";
 
-import type { Repository } from "commons/graphql";
+import type { RepositoryProps } from "commons/graphql";
 import githubGraphQL, { PROJECTS, CONTRIBUTIONS } from "commons/graphql";
 import { Page } from "components/pages";
 import Sections, { Contributions, Me, Projects } from "sections";
 import Nav from "sections/nav";
 
 interface HomeProps {
-    projects: Repository[];
-    contributions: Repository[];
+    projects: RepositoryProps[];
+    contributions: RepositoryProps[];
 }
 
 export default function Home({ projects, contributions }: HomeProps) {
@@ -33,10 +33,10 @@ export const getStaticProps: GetStaticProps<HomeProps> = async () => {
     /* eslint-disable @typescript-eslint/no-unsafe-assignment */
     const projectsResp = await githubGraphQL(PROJECTS);
     const projectsJson = await projectsResp.json();
-    const projects: Repository[] = projectsJson.data.viewer.pinnedItems.nodes;
+    const projects: RepositoryProps[] = projectsJson.data.viewer.pinnedItems.nodes;
     const contributionsResp = await githubGraphQL(CONTRIBUTIONS);
     const contributionsJson = await contributionsResp.json();
-    const contributions: Repository[] =
+    const contributions: RepositoryProps[] =
         contributionsJson.data.viewer.repositoriesContributedTo.nodes;
 
     return {
