+++ src/sections/me/index.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import { forwardRef } from "react";
 
 import { GITHUB } from "commons/config";
-import { Header } from "components";
+import { Header } from "components/common";
 import Lanyard from "sections/me/lanyard";
 import Socials from "sections/me/socials";
