--- src/sections/projects/project/details/index.ts
@@ -1,2 +1,0 @@
-export { Forks } from "sections/projects/project/details/Forks";
-export { Stars } from "sections/projects/project/details/Stars";
